requirements.txt
flatbuffers
pytest
psycopg2-binary